fix: skip fluent-schema test with node < 8
